pyproject.toml
[project]
name = "cronctl"
version = "0.1.0"
description = "Lightweight, AI-agent-compatible cron job manager for local environments"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["cron", "scheduler", "cli", "mcp", "ai-agent", "automation"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: System :: Systems Administration",
  "Topic :: Utilities",
  "Typing :: Typed",
]

dependencies = [
  "click>=8.1",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
mcp = ["mcp>=1.0"]
notify = ["httpx>=0.27"]
all = ["cronctl[mcp,notify]"]
dev = [
  "mypy>=1.10",
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.4",
  "types-PyYAML",
]

[project.scripts]
cronctl = "cronctl.__main__:main"

[project.urls]
Homepage = "https://github.com/yourname/cronctl"
Repository = "https://github.com/yourname/cronctl"
Issues = "https://github.com/yourname/cronctl/issues"
Documentation = "https://github.com/yourname/cronctl/tree/main/docs"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cronctl"]
include = ["src/cronctl/skill/SKILL.md"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM", "TCH"]
ignore = ["UP017", "UP042"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
